refactor: healthz tests now live in healthz package
