+++ main_test.go
@@ -17,35 +17,10 @@ import (
 	"url-short/internal/database"
 	"url-short/internal/test/fixtures"
 	testHelpers "url-short/internal/test/helpers"
-	"url-short/internal/transport/http/health"
 	httpHelpers "url-short/internal/transport/http/helper"
 	"url-short/internal/transport/http/shorturls"
 	"url-short/internal/transport/http/users"
 )
-
-func TestHealthEndpoint(t *testing.T) {
-	t.Run("test healthz endpoint", func(t *testing.T) {
-		request, _ := http.NewRequest(http.MethodGet, "/api/v1/healthz", nil)
-		response := httptest.NewRecorder()
-
-		health.GetHealth(response, request)
-
-		got := health.GetHealthHTTPResponseBody{}
-		err := json.NewDecoder(response.Body).Decode(&got)
-
-		if err != nil {
-			t.Errorf("unable to parse response %q into %q", response.Body, got)
-		}
-
-		if got.Status != "ok" {
-			t.Errorf("status field must be okay on health response got %q wanted %q", got.Status, "ok")
-		}
-
-		if response.Result().StatusCode != http.StatusOK {
-			t.Error("endpoint must return 200")
-		}
-	})
-}
 
 func TestPostUser(t *testing.T) {
 	dbURL := os.Getenv("PG_CONN")
